(Jackie) Placed Card.js in CharacterIndex.js file so that cards do not appear on login/signup page.
